+++ packages/react-query-next-experimental/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tanstack/react-query-next-experimental
 
+## 5.100.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.100.1
+
+## 5.100.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.100.0
+
 ## 5.99.2
 
 ### Patch Changes